Properly working to draw each precinct separately
However, because each precinct is in it's own layer, this is taking a huge toll on the graphic drawing. There may be a way to condense multiple sources into layers so long as those precincts are not touching the color gap shouldn't matter. HOWEVER, this will come at the cost of individually accessing the precincts layer so not sure.